Another trick to exploit one bit flip.
We need to patch __sys_setuid as it already call prepare_cred() -> commit_creds(), we need to make it prepare_kernel_creds() -> commit_creds()
We need to change it from physmap. The starting address when KASLR is off = 0xffff880000000000.
You just flip the the U bit in the page table corresponding to __sys_setuid (inside the physmap) and change the page table of that into W and U.
Now you have full control over __sys_setuid.