Two kernel challenges that I created for 3kCTF21.
echo:
For echo the vulnerability is easy to find, a 1 line syscall that only do copy_user_generic_unrolled(src, dst, 8) note that src and dst are both controlled.
What we can do is to bruteforce every place that can hold modprobe_path and change that to your shell script. (refer to the exploit)

klibrary:
This challenge is a vulnerable loadable kernel module (race condition)
The locks are not properly placed, and userfault_fd is enabled in the kernel (you can use it for reliable exploit and you're free not to use it.
The target struct for this challenge was tty_struct, overwrite one for arbitrary read and the other for arbitrary write.
Search for the cred struct of the exploit binary using the arbitrary read and patch its cred using arbitrary write (refer to the exploit for more details).
Note- if SMAP wasn't enabled we were be able to pivot the stack and execute ROP but SMAP was enabled to introduce the arb read/write attack.